Rollback if deployed new version is not working

+++ upgrade.sh
@@ -5,7 +5,12 @@ set -eo pipefail
 export COMPOSE_DOCKER_CLI_BUILD=1
 export DOCKER_BUILDKIT=1
 
+IMAGE_NAME='ghcr.io/mastodon/mastodon'
+
 sudo setfacl -R -m "user:$USER:rx" postgres14
+
+# Save image
+docker tag $IMAGE_NAME:latest $IMAGE_NAME:stable
 
 docker-compose build # --parallel
 
@@ -14,8 +19,15 @@ docker-compose run --rm -e SKIP_POST_DEPLOYMENT_MIGRATIONS=true web rails db:mig
 # docker-compose down
 docker-compose up -d
 
-curl -fs https://parfait.day/health
+if curl -fso /dev/null https://parfait.day/api/v1/instance; then
+  docker-compose run --rm web rails db:migrate
+  docker-compose run --rm web tootctl cache clear
+  docker image prune -f
+else
+  # Rollback
+  docker tag $IMAGE_NAME:stable $IMAGE_NAME:latest
+  docker-compose up -d
+  exit 1
+fi
 
-docker-compose run --rm web rails db:migrate
-docker-compose run --rm web tootctl cache clear
-docker image prune -f
+echo "OK"
